test(go): drive every indicator through batch, and fix what that found
The generated golden suite replayed the shared input through `update` and
checked it against the Rust fixtures. `batch` was never exercised, so a batch
that disagreed with streaming -- or that did not exist -- passed unnoticed for all
514. It now runs both passes against the same fixtures.

Three archetypes need a different comparison, and saying so is the point rather
than a wrinkle: a bar builder completes an unpredictable number of bars per
candle, so its batch is the concatenation and cannot be split back into rows;
and Footprint reports the whole book after each trade, so its batch holds the
final snapshot, which is the fixture's last row.

511 of the 514 passed immediately. The three that did not were all on the warmup
row, and all the same mistake: a row the indicator did not produce came back as
zeros, which is indistinguishable from a value it did produce.

`LeadLagCrossCorrelation` is the one multi-output struct with an integer field,
and an integer has no NaN, so its lag read back as 0 while streaming reports the
whole row as absent. I had written that down as a caveat when adding the
multi-output batch; the test showed the caveat was the defect. Integers surface
as `f64` in the C output structs now, so every field of every one can carry NaN
-- the lag is a bar offset and is exact in an `f64`.

`TpoProfile` and `VolumeProfile` NaN-filled their values block but left the
scalar array untouched, so those fields read back as zeros.

A spot-check would likely have skipped all three: they are on row 0 of three
indicators out of 514.

With the WASM batch suite added earlier, two independent language paths now
cover every indicator's batch. C#, Java and R keep their per-shape tests, which
is where their own marshalling lives.

Generator updated in ScriptHelpers f3585ed. Rust chain green, clippy clean, Go
green with 514 batch subtests, C# 565, Java 558, R 120010 assertions.

+++ bindings/go/include/wickra.h
@@ -1348,7 +1348,7 @@ typedef struct WickraKstOutput {
 } WickraKstOutput;
 
 typedef struct WickraLeadLagCrossCorrelationOutput {
-    int64_t lag;
+    double lag;
     double correlation;
 } WickraLeadLagCrossCorrelationOutput;
